Breaking integration tests into individual packages to improve visibility into timing and issues
