+++ internal/test/integration/accumulate/accumulate_test.go
@@ -1,4 +1,4 @@
-package integration
+package accumulate
 
 import (
 	"context"
