+++ internal/test/integration/flatmap/map_error_test.go
@@ -1,4 +1,4 @@
-package integration
+package flatmap
 
 import (
 	"context"
