+++ internal/test/integration/reduce/reduce_test.go
@@ -1,9 +1,7 @@
-package integration
+package reduce
 
 import (
 	"context"
-	"fmt"
-	"strings"
 	"testing"
 
 	"github.com/go-sif/sif"
@@ -17,7 +15,7 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func createTestCollectDataFrame(t *testing.T, numRows int) sif.DataFrame {
+func createTestReduceDataFrame(t *testing.T, numRows int) sif.DataFrame {
 	row := []byte("{\"col1\": \"abc\"}")
 	data := make([][]byte, numRows)
 	for i := 0; i < len(data); i++ {
@@ -34,38 +32,47 @@ func createTestCollectDataFrame(t *testing.T, numRows int) sif.DataFrame {
 	return dataframe
 }
 
-func TestCollect(t *testing.T) {
+func TestReduce(t *testing.T) {
 	// create dataframe
-	frame, err := createTestCollectDataFrame(t, 10).To(
-		ops.AddColumn("res", &sif.VarStringColumnType{}),
+	numRows := 100
+	frame, err := createTestReduceDataFrame(t, numRows).To(
+		ops.AddColumn("count", &sif.Uint32ColumnType{}),
 		ops.Map(func(row sif.Row) error {
-			col1, err := row.GetString("col1")
-			if err != nil {
-				return err
-			}
-			err = row.SetVarString("res", fmt.Sprintf("%s", strings.ToUpper(col1)))
+			err := row.SetInt32("count", int32(1))
 			if err != nil {
 				return err
 			}
 			return nil
 		}),
 		ops.RemoveColumn("col1"),
-		util.Collect(2), // 2 partitions because there are 10 rows and 5 per partition
+		ops.Reduce(func(row sif.Row) ([]byte, error) {
+			return []byte{byte(1)}, nil
+		}, func(lrow sif.Row, rrow sif.Row) error {
+			lval, err := lrow.GetInt32("count")
+			if err != nil {
+				return err
+			}
+			rval, err := rrow.GetInt32("count")
+			if err != nil {
+				return err
+			}
+			return lrow.SetInt32("count", lval+rval)
+		}),
+		util.Collect(10), // should be 1 partitions because we are summing to a single row, but collect extra as a test
 	)
 	require.Nil(t, err)
 
-	// run dataframe
+	// run dataframe and verify results
 	res, err := siftest.LocalRunFrame(context.Background(), frame, &cluster.NodeOptions{}, 2)
 	require.Nil(t, err)
 	require.NotNil(t, res)
-	require.NotNil(t, res.Collected)
+	require.Equal(t, 1, len(res.Collected))
 	for _, part := range res.Collected {
-		part.ForEachRow(func(row sif.Row) error {
-			val, err := row.GetVarString("res")
-			require.Nil(t, err)
-			require.Equal(t, "ABC", val)
-			return nil
-		})
+		require.Equal(t, 1, part.GetNumRows())
+		row := part.GetRow(0)
+		count, err := row.GetInt32("count")
+		require.Nil(t, err)
+		require.EqualValues(t, numRows, count)
+		break
 	}
-	require.Nil(t, err)
 }
