+++ internal/test/integration/repartition/repartition_test.go
@@ -1,4 +1,4 @@
-package integration
+package repartition
 
 import (
 	"context"
